# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bracket Golf API - U.S. Amateur Bracket Challenge
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brackets:
    get:
      operationId: getBrackets
      tags: [brackets]
      summary: Get all brackets for the current user
      responses:
        "200":
          description: User's brackets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bracket"
    post:
      operationId: createBracket
      tags: [brackets]
      summary: Create a new bracket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BracketInput"
      responses:
        "201":
          description: Created bracket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bracket"
  /brackets/{bracketId}:
    get:
      operationId: getBracket
      tags: [brackets]
      summary: Get a specific bracket with picks
      parameters:
        - name: bracketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bracket with picks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BracketWithPicks"
        "404":
          description: Bracket not found
    delete:
      operationId: deleteBracket
      tags: [brackets]
      summary: Delete a bracket
      parameters:
        - name: bracketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /brackets/{bracketId}/submit:
    post:
      operationId: submitBracket
      tags: [brackets]
      summary: Submit a completed bracket
      parameters:
        - name: bracketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submitted bracket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bracket"

  # --- PICKS ---
  /brackets/{bracketId}/picks:
    get:
      operationId: getBracketPicks
      tags: [picks]
      summary: Get all picks for a bracket
      parameters:
        - name: bracketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bracket picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BracketPick"
    put:
      operationId: savePick
      tags: [picks]
      summary: Save or update a single pick (autosave)
      parameters:
        - name: bracketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickInput"
      responses:
        "200":
          description: Updated picks for this bracket
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BracketPick"

  # --- LEADERBOARD ---
...
components:
  schemas:
...
    BracketInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
    Bracket:
      type: object
      required: [id, userId, tournamentId, name, submitted, score, maxPossibleScore, totalPicks, completedPicks, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        tournamentId:
          type: string
        name:
          type: string
        submitted:
          type: boolean
        submittedAt:
          type: ["string", "null"]
        score:
          type: number
        maxPossibleScore:
          type: number
        rank:
          type: ["integer", "null"]
        totalPicks:
          type: number
        completedPicks:
          type: number
        championGolferId:
          type: ["string", "null"]
        championName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BracketWithPicks:
      type: object
      required: [id, userId, tournamentId, name, submitted, score, maxPossibleScore, totalPicks, completedPicks, picks, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        tournamentId:
          type: string
        name:
          type: string
        submitted:
          type: boolean
        submittedAt:
          type: ["string", "null"]
        score:
          type: number
        maxPossibleScore:
          type: number
        rank:
          type: ["integer", "null"]
        totalPicks:
          type: number
        completedPicks:
          type: number
        championGolferId:
          type: ["string", "null"]
        championName:
          type: ["string", "null"]
        picks:
          type: array
          items:
            $ref: "#/components/schemas/BracketPick"
        createdAt:
          type: string
        updatedAt:
          type: string

    # --- BRACKET PICK ---
    PickInput:
      type: object
      required: [matchupId, selectedGolferId]
      properties:
        matchupId:
          type: string
        selectedGolferId:
          type: string
    BracketPick:
      type: object
      required: [id, bracketId, matchupId, selectedGolferId, round, roundNumber, pointsPossible, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        bracketId:
          type: string
        matchupId:
          type: string
        selectedGolferId:
          type: string
        round:
          type: string
          enum: [R64, R32, R16, QF, SF, F]
        roundNumber:
          type: number
        pointsPossible:
          type: number
        status:
          type: string
          enum: [pending, correct, incorrect, eliminated]
        isCorrect:
          type: ["boolean", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    # --- LEADERBOARD ---
...