# lib/api-zod/src/generated/types/bracketPick.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bracket Golf API - U.S. Amateur Bracket Challenge
 * OpenAPI spec version: 0.1.0
 */
import type { BracketPickRound } from './bracketPickRound';
import type { BracketPickStatus } from './bracketPickStatus';

export interface BracketPick {
  id: string;
  bracketId: string;
  matchupId: string;
  selectedGolferId: string;
  round: BracketPickRound;
  roundNumber: number;
  pointsPossible: number;
  status: BracketPickStatus;
  /** @nullable */
  isCorrect?: boolean | null;
  createdAt: string;
  updatedAt: string;
}